requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & ML
openai==1.10.0
pillow==10.2.0
pdf2image==1.17.0
pytesseract==0.3.10

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
python-multipart==0.0.6
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.26.0

# Validation & Security
email-validator==2.1.0
python-magic==0.4.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Logging & Monitoring
loguru==0.7.2

# Date/Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# File handling
PyPDF2==3.0.1
aiosqlite==0.19.0
greenlet==3.0.3
